WorldNews/requirements.txt
beautifulsoup4
urllib3
requests
pandas
numpy
Pillow
nltk
fsspec
